% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd.R
\name{as.character.zdd}
\alias{as.character.zdd}
\title{zdd as character}
\usage{
\method{as.character}{zdd}(x, ...)
}
\arguments{
\item{x}{a zdd}

\item{...}{all the other stuff}
}
\value{
the string ID of the zdd node
}
\description{
zdd as character
}
\examples{
as.character( as_zdd(3L) )
}
