% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd.R
\name{as.integer.zdd}
\alias{as.integer.zdd}
\title{zdd as integer}
\usage{
\method{as.integer}{zdd}(x, ...)
}
\arguments{
\item{x}{a zdd}

\item{...}{all the other stuff}
}
\value{
the integer value of the zdd node
}
\description{
zdd as integer
}
\examples{
as.integer( as_zdd(3L) )
as.integer( zdd0() )
}
