% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_zdd.R
\name{as_zdd}
\alias{as_zdd}
\title{as_zdd}
\usage{
as_zdd(x)
}
\arguments{
\item{x}{the integer value of the node to be added or the string hash of the node to pull from the store}
}
\value{
the single variable elementary zdd (the so-called "set containing \code{value}")
}
\description{
as_zdd
}
\examples{
as_zdd(3L)
as_zdd('d7281766b28bdc9025520edeeffecdd9')
#as_zdd('this one does not exist')  # this one throws an error
}
