% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutsets.R
\name{cutsets}
\alias{cutsets}
\title{cutsets}
\usage{
cutsets(zdd)
}
\arguments{
\item{zdd}{a zdd list object}
}
\value{
a list of integer vectors representing a family of sets (cutsets)
}
\description{
cutsets
}
\examples{
cutsets( zdd_union(as_zdd(3L), as_zdd(4L) ) )
cutsets( zdd_union(as_zdd(4L), as_zdd(3L) ) )
cutsets( zdd_union(as_zdd(4L), zdd0()     ) )
cutsets( zdd_union(as_zdd(4L), zdd1()     ) )
}
