% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zdd.R
\name{get_zdd}
\alias{get_zdd}
\title{get_zdd}
\usage{
get_zdd(hash)
}
\arguments{
\item{hash}{name of the zdd in the zdd_store}
}
\value{
zdd list object
}
\description{
get_zdd
}
\examples{
get_zdd(as_zdd(3L)$hash)
}
