% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_anot.R
\name{\%anot\%}
\alias{\%anot\%}
\title{infix operator for zdd_anot}
\usage{
a \%anot\% b
}
\arguments{
\item{a}{a zdd list object}

\item{b}{a zdd list object}
}
\value{
zdd_anot(a,b)
}
\description{
infix operator for zdd_anot
}
