% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_equal.R
\name{is_equal}
\alias{is_equal}
\title{is_equal}
\usage{
is_equal(zddP, zddQ)
}
\arguments{
\item{zddP}{a zdd in the zdd_store}

\item{zddQ}{a zdd in the zdd_store}
}
\value{
TRUE if the input zdd's are equal
}
\description{
is_equal
}
\examples{
is_equal(as_zdd(1L), as_zdd(1L)) # TRUE
}
