% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one.R
\name{is_one}
\alias{is_one}
\title{is_one}
\usage{
is_one(zdd)
}
\arguments{
\item{zdd}{a zdd list object}
}
\value{
TRUE if the zdd matches the output of zdd1()
}
\description{
is_one
}
\examples{
is_one(zdd1())     # TRUE
is_one(zdd(1L))    # FALSE
}
