% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_zero.R
\name{is_zero}
\alias{is_zero}
\title{is_zero}
\usage{
is_zero(zdd)
}
\arguments{
\item{zdd}{a zdd list object}
}
\value{
TRUE if the zdd matches the output of zdd0()
}
\description{
is_zero
}
\examples{
is_zero(zdd0())     # TRUE
is_zero(zdd(0L)) # FALSE
}
