% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd.R
\name{print.zdd}
\alias{print.zdd}
\title{print zdd}
\usage{
\method{print}{zdd}(x, ...)
}
\arguments{
\item{x}{a zdd}

\item{...}{all the other stuff}
}
\value{
printed zdd
}
\description{
print zdd
}
\examples{
print(zdd0())
print(zdd(3L))
}
