% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect.R
\name{protect}
\alias{protect}
\alias{protect.zdd}
\title{protect}
\usage{
protect(x)

\method{protect}{zdd}(x)
}
\arguments{
\item{x}{the ZDD to protect, including subnodes}
}
\value{
the ZDD, protected
}
\description{
protect
}
\examples{
a <- as_zdd(list(c(1,2,3), 4, c(5,6)))
protect(a)
reset_zdd_store(keep_protected = TRUE) # preserves the protected ZDD assigned to `a`
}
