% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_zdd_store.R
\name{reset_zdd_store}
\alias{reset_zdd_store}
\title{reset_zdd_store}
\usage{
reset_zdd_store(keep_protected = TRUE, quiet = FALSE)
}
\arguments{
\item{keep_protected}{when TRUE, preserves the ZDD nodes that are marked protected}

\item{quiet}{when FALSE, print a message whenever this command is run}
}
\description{
reset_zdd_store
}
\examples{
reset_zdd_store(quiet = TRUE)
}
