% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_and.R
\name{zdd_and}
\alias{zdd_and}
\title{zdd_and}
\usage{
zdd_and(...)
}
\arguments{
\item{...}{a list of objects that can be coerced to zdds}
}
\value{
a zdd that is the crossmultiplication of all inputs
}
\description{
zdd_and
}
\examples{
zdd_and(1,2,3,4)
}
