% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_anot.R
\name{zdd_anot}
\alias{zdd_anot}
\title{zdd_anot}
\usage{
zdd_anot(zddP, zddQ)
}
\arguments{
\item{zddP}{a zdd list object}

\item{zddQ}{a zdd list object}
}
\value{
a zdd list object that is the elements of the first object that are nonminimal to the second
}
\description{
zdd_anot
}
\examples{
cutsets( (zdd(3) | zdd(4)) \%anot\% zdd(4))
cutsets( (zdd(3) * zdd(4)) \%anot\% zdd(4))
}
