% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd.R
\name{zdd_binary_function}
\alias{zdd_binary_function}
\title{helper function to store the results of binary operations}
\usage{
zdd_binary_function(P, op, Q)
}
\arguments{
\item{P}{first zdd argument}

\item{op}{binary function}

\item{Q}{second zdd argument}
}
\description{
helper function to store the results of binary operations
}
