% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_crossproduct.R
\name{zdd_crossproduct}
\alias{zdd_crossproduct}
\title{zdd_crossproduct}
\usage{
zdd_crossproduct(zddP, zddQ)
}
\arguments{
\item{zddP}{a zdd list object}

\item{zddQ}{a zdd list object}
}
\value{
a zdd list object that is the crossproduct of the two input zdds
}
\description{
zdd_crossproduct
}
\examples{
zdd_crossproduct(as_zdd(2L), as_zdd(3L))
as_zdd(2L) * as_zdd(3L)
(as_zdd(1L)|as_zdd(2L)|as_zdd(3L)) * (as_zdd(3L)|as_zdd(4L)|as_zdd(5L))
}
