% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_difference.R
\name{zdd_difference}
\alias{zdd_difference}
\title{zdd_difference}
\usage{
zdd_difference(zddP, zddQ)
}
\arguments{
\item{zddP}{a zdd list object}

\item{zddQ}{a zdd list object}
}
\value{
a zdd list object that is the difference between the two zdds
}
\description{
zdd_difference
}
\examples{
zdd_union(as_zdd(2L), as_zdd(3L))
as_zdd(2L) | as_zdd(3L) - 3L
}
