% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zdd_fxns}
\alias{zdd_fxns}
\title{Store for ZDD operations}
\format{
An object of class \code{environment} of length 0.
}
\usage{
zdd_fxns
}
\description{
This environment stores results of binary operations between ZDDs
}
\keyword{datasets}
