% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd.R
\name{zdd_hash}
\alias{zdd_hash}
\title{zdd hash}
\usage{
zdd_hash(value, p0, p1)
}
\arguments{
\item{value}{the integer value of the node}

\item{p0}{the negative compliment of the zdd with respect to \code{value}, expressed as class zdd}

\item{p1}{the positive compliment of the zdd with respect to \code{value}, expressed as class zdd}
}
\value{
hash
}
\description{
zdd hash
}
\examples{
zddr:::zdd_hash(1L, zdd0(), zdd1())
}
