% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_minimize.R
\name{zdd_minimize}
\alias{zdd_minimize}
\title{zdd_minimize}
\usage{
zdd_minimize(zdd, Q = NULL)
}
\arguments{
\item{zdd}{a zdd object}

\item{Q}{a NULL object - nothing is done with this input, a refactor should be performed at some point to remove it}
}
\value{
a zdd object that contains only minimal solutions
}
\description{
zdd_minimize
}
