% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_or.R
\name{zdd_or}
\alias{zdd_or}
\title{zdd_or}
\usage{
zdd_or(...)
}
\arguments{
\item{...}{a list of objects that can be coerced to zdds}
}
\value{
a zdd that is the union of all inputs
}
\description{
zdd_or
}
\examples{
zdd_or(1,2,3,4)
zdd_or( zdd_and(1,2), zdd_and(3,4) ) # { {1,2}, {3,4} }
zdd_or( zdd_and(1,2,3,4,5), 4)       # { {4} }
}
