% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zdd_store}
\alias{zdd_store}
\title{Store for Zero-suppressed BDD nodes}
\format{
An object of class \code{environment} of length 0.
}
\usage{
zdd_store
}
\description{
This environment stores objects of type 'zdd'
}
\keyword{datasets}
