% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_subset0.R
\name{zdd_subset0}
\alias{zdd_subset0}
\title{zdd_subset0}
\usage{
zdd_subset0(zdd, value)
}
\arguments{
\item{zdd}{a zdd list object}

\item{value}{a value}
}
\value{
the subset of zdd such as value = TRUE
}
\description{
zdd_subset0
}
\examples{
zdd_subset0(zdd(2L), 2L)# ONE
zdd_subset0(zdd(2L), 1L)# {2}
zdd_subset0(zdd(1L), 2L)# ZERO
zdd_subset0(zdd_and(3,4,5), 2L)   # {3,4,5}
zdd_subset0(zdd_and(3,4,5), 5L)   # {}
}
