% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_union.R
\name{zdd_union}
\alias{zdd_union}
\title{zdd_union}
\usage{
zdd_union(zddP, zddQ)
}
\arguments{
\item{zddP}{a zdd list object}

\item{zddQ}{a zdd list object}
}
\value{
a zdd list object that is the union of the two input zdds
}
\description{
zdd_union
}
\examples{
zdd_union(as_zdd(2L), as_zdd(3L))
}
