% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_zdd.R
\name{as_zdd}
\alias{as_zdd}
\title{as_zdd}
\usage{
as_zdd(value)
}
\arguments{
\item{value}{the integer value of the node to be added}
}
\value{
the single variable elementary zdd (the so-called "set containing \code{value}")
}
\description{
as_zdd
}
\examples{
as_zdd(3L)
}
