% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutsets.R
\name{cutsets}
\alias{cutsets}
\title{cutsets}
\usage{
cutsets(zdd)
}
\arguments{
\item{zdd}{a zdd list object}
}
\value{
a list of integer vectors representing a family of sets (cutsets)
}
\description{
cutsets
}
\examples{
library(magrittr)
zdd3 <- as_zdd(3L)
zdd4 <- as_zdd(4L)
zdd3 \%>\% zdd_union(zdd4) \%>\% cutsets()
zdd4 \%>\% zdd_union(zdd3) \%>\% cutsets()
zdd4 \%>\% zdd_union(zdd0()) \%>\% cutsets()
zdd4 \%>\% zdd_union(zdd1()) \%>\% cutsets()
}
