% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutsets.R
\name{cutsets}
\alias{cutsets}
\title{cutsets}
\usage{
cutsets(zdd)
}
\arguments{
\item{zdd}{a zdd list object}
}
\value{
a list of integer vectors representing a family of sets (cutsets)
}
\description{
cutsets
}
\examples{
cutsets( zdd(3) | zdd(4) )
cutsets( zdd(4) | zdd(3) )
cutsets( zdd(4) | FALSE  )
cutsets( zdd(4) | TRUE   )
}
