% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_crossproduct.R
\name{zdd_crossproduct}
\alias{zdd_crossproduct}
\title{zdd_crossproduct}
\usage{
zdd_crossproduct(zddP, zddQ)
}
\arguments{
\item{zddP}{a zdd list object}

\item{zddQ}{a zdd list object}
}
\value{
a zdd list object that is the crossproduct of the two input zdds
}
\description{
zdd_crossproduct
}
\examples{
zdd_crossproduct(zdd(2L), zdd(3L))
zdd(2L) * zdd(3L)
(zdd(1L)|zdd(2L)|zdd(3L)) * (zdd(3L)|zdd(4L)|zdd(5L))
}
