% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_minimize.R
\name{zdd_minimize}
\alias{zdd_minimize}
\title{zdd_minimize}
\usage{
zdd_minimize(zdd)
}
\arguments{
\item{zdd}{a zdd object}
}
\value{
a zdd object that has only minimal solutions
}
\description{
zdd_minimize
}
\examples{
zdd(2L) | zdd(3L) & zdd(3L)   # { {2}, {3} }
(zdd(2L) | zdd(3L)) & zdd(3L) # {   {3}    }
}
