% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdd_subset0.R
\name{zdd_subset0}
\alias{zdd_subset0}
\title{zdd_subset0}
\usage{
zdd_subset0(zdd, value)
}
\arguments{
\item{zdd}{a zdd list object}

\item{value}{a value}
}
\value{
the subset of zdd such as value = TRUE
}
\description{
zdd_subset0
}
\examples{
zdd_subset0(as_zdd(2L), 2L)# ONE
zdd_subset0(as_zdd(2L), 1L)# {2}
zdd_subset0(as_zdd(1L), 2L)# ZERO
library(magrittr)
zdd345 <-
  zdd(3L, zdd0(), zdd1()) \%>\%
  zdd(4L, zdd0(), .     ) \%>\%
  zdd(5L, zdd0(), .     )
zdd_subset0(zdd345, 4L)   # {3,5}
}
